refactor: Extract ProcessClock into its own type
